refactor(tests): extract response test vectors, sanitization, and URL validation into dedicated test modules
Extract shared XSS test vectors into _response_test_vectors.py.
Move sanitization tests to test_sanitization.py.
Move URL validation tests to test_url_validation.py.
Rename test_responses_models.py → test_responses.py (reduced scope).

Refs: #459
